Revert "Merge branch 'master' of https://github.com/frappe/frappe into frappe-master"
This reverts commit 9475df0ea0b21e64913f1b35f159004390051857, reversing
changes made to 0260b518ff5f613e9adf4c09da004981db49fa36.

# Conflicts:
#	frappe/translations/es-AR.csv

+++ frappe/build.py
@@ -4,7 +4,6 @@
 from __future__ import unicode_literals, print_function
 from frappe.utils.minify import JavascriptMinify
 import subprocess
-import warnings
 
 from six import iteritems, text_type
 
@@ -26,12 +25,12 @@ def setup():
 		except ImportError: pass
 	app_paths = [os.path.dirname(pymodule.__file__) for pymodule in pymodules]
 
-def bundle(no_compress, make_copy=False, restore=False, verbose=False):
+def bundle(no_compress, make_copy=False, verbose=False):
 	"""concat / minify js files"""
 	# build js files
 	setup()
 
-	make_asset_dirs(make_copy=make_copy, restore=restore)
+	make_asset_dirs(make_copy=make_copy)
 
 	# new nodejs build system
 	command = 'node --use_strict ../apps/frappe/frappe/build.js --build'
@@ -61,8 +60,7 @@ def watch(no_compress):
 
 	# 	time.sleep(3)
 
-def make_asset_dirs(make_copy=False, restore=False):
-	# don't even think of making assets_path absolute - rm -rf ahead.
+def make_asset_dirs(make_copy=False):
 	assets_path = os.path.join(frappe.local.sites_path, "assets")
 	for dir_path in [
 			os.path.join(assets_path, 'js'),
@@ -82,28 +80,11 @@ def make_asset_dirs(make_copy=False, restore=False):
 
 		for source, target in symlinks:
 			source = os.path.abspath(source)
-			if os.path.exists(source):
-				if restore:
-					if os.path.exists(target):
-						if os.path.islink(target):
-							os.unlink(target)
-						else:
-							shutil.rmtree(target)
-						shutil.copytree(source, target)
-				elif make_copy:
-					if os.path.exists(target):
-						warnings.warn('Target {target} already exists.'.format(target = target))
-					else:
-						shutil.copytree(source, target)
+			if not os.path.exists(target) and os.path.exists(source):
+				if make_copy:
+					shutil.copytree(source, target)
 				else:
-					if os.path.exists(target):
-						if os.path.islink(target):
-							os.unlink(target)
-						else:
-							shutil.rmtree(target)
 					os.symlink(source, target)
-			else:
-				warnings.warn('Source {source} does not exists.'.format(source = source))
 
 def build(no_compress=False, verbose=False):
 	assets_path = os.path.join(frappe.local.sites_path, "assets")
